Move render_node.py
